requirements.txt
django-storages
boto3
django-tailwind==3.6.0
Django>=4.2,<5.0
gunicorn>=20.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
whitenoise>=6.0.0
dj-database-url>=1.0.0
azure-communication-email
django-ckeditor
stripe
Pillow
django-formtools
django-sslserver
# uszipcode removed after migrating to local ZipCode dataset (seeded via full load command)
sqlalchemy-mate<2.0
Faker
